P3-50(#40) [FEAT]: 매수/매도 주문 매칭

+++ order-service/src/main/java/finpago/orderservice/order/dto/OrderCreateReqDto.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class OrderCreateReqDto {
-    private Float offerQuantity;
+    private Long offerQuantity;
     private Long offerPrice;
     private String stockTicker;
     private String offerType; // "BUY" or "SELL"
